requirements.txt
playwright>=1.42.0
requests>=2.31.0
python-dotenv>=1.0.0
azure-storage-blob>=12.19.1
tqdm>=4.66.2
